checkpoint in port to macOS
Attempting to port code to macOS.

The rogomatic tool will crash with a "segmentation fault" after a few seconds.
This problem needs to be debugged.

It is also possible that a number of the compiler warnings, rather than being
silenced via "-Wno-\*" flags, need to be addressed and fixed.

XXX - This is a port in progress - XXX - Pull requests to fix are welcome!